test: extract TestUtils
