+++ packages/effect/test/unstable/ai/McpServer/McpConformance/CompletionTest.ts
@@ -3,7 +3,7 @@ import * as Effect from "effect/Effect"
 import * as Schema from "effect/Schema"
 import type * as McpProtocol from "effect/unstable/ai/McpProtocol"
 import * as McpSchema from "effect/unstable/ai/McpSchema"
-import { McpConformanceTest, type TestLayer } from "./McpConformanceTest.ts"
+import { McpConformance, type McpConformanceLayer } from "./McpConformance.ts"
 
 const decodeCompletion = Schema.decodeUnknownEffect(McpSchema.CompleteResult)
 
@@ -15,7 +15,7 @@ const complete = (
   argument: { readonly name: string; readonly value: string }
 ) =>
   Effect.gen(function*() {
-    const test = yield* McpConformanceTest
+    const test = yield* McpConformance
     const initialized = yield* test.initialize({ server: "features" })
     yield* test.notifyInitialized(initialized)
     const response = yield* test.send(initialized, {
@@ -29,22 +29,20 @@ const complete = (
     )
   })
 
-export const suite = (protocol: McpProtocol.ProtocolAdapter, layer: TestLayer) =>
+export const suite = (protocol: McpProtocol.ProtocolAdapter, layer: McpConformanceLayer) =>
   it.layer(layer)(`Mcp Conformance (${protocol.protocolVersion})`, (it) => {
     describe("Completion", () => {
       // Shared by the 2024-11-05, 2025-03-26, and 2025-06-18 specifications,
       // except completion context, which was added in 2025-06-18.
-      if (protocol.protocolVersion !== "2024-11-05") {
-        describe("Capabilities", () => {
-          it.effect("MUST advertise completions when argument completion is supported", () =>
-            Effect.gen(function*() {
-              const test = yield* McpConformanceTest
-              const initialized = yield* test.initialize({ server: "features" })
+      describe("Capabilities", () => {
+        it.effect("MUST advertise completions when argument completion is supported", () =>
+          Effect.gen(function*() {
+            const test = yield* McpConformance
+            const initialized = yield* test.initialize({ server: "features" })
 
-              assert.property(initialized.message.result.capabilities, "completions")
-            }))
-        })
-      }
+            assert.property(initialized.message.result.capabilities, "completions")
+          }))
+      })
 
       describe("Requesting Completions", () => {
         it.effect("MUST complete a prompt argument", () =>
@@ -66,46 +64,44 @@ export const suite = (protocol: McpProtocol.ProtocolAdapter, layer: TestLayer) =
 
             assert.deepStrictEqual(result.completion.values, ["alpha", "beta"])
           }))
-        if (protocol.protocolVersion === "2025-06-18") {
-          // FIX: The public completion-handler API currently receives only the
-          // argument value and drops the June `context.arguments` object.
-          it.effect.skip("MUST pass previously resolved argument context to the completion handler", () =>
-            Effect.gen(function*() {
-              const test = yield* McpConformanceTest
-              const initialized = yield* test.initialize({ server: "features" })
-              yield* test.notifyInitialized(initialized)
-              const response = yield* test.send(initialized, {
-                jsonrpc: "2.0",
-                id: 2,
-                method: "completion/complete",
-                params: {
-                  ref: {
-                    type: "ref/prompt",
-                    name: "ContextCompletionPrompt"
-                  },
-                  argument: {
-                    name: "value",
-                    value: "c"
-                  },
-                  context: {
-                    arguments: {
-                      locale: "en"
-                    }
+        // FIX: The public completion-handler API currently receives only the
+        // argument value and drops the June `context.arguments` object.
+        it.effect.skip("MUST pass previously resolved argument context to the completion handler", () =>
+          Effect.gen(function*() {
+            const test = yield* McpConformance
+            const initialized = yield* test.initialize({ server: "features" })
+            yield* test.notifyInitialized(initialized)
+            const response = yield* test.send(initialized, {
+              jsonrpc: "2.0",
+              id: 2,
+              method: "completion/complete",
+              params: {
+                ref: {
+                  type: "ref/prompt",
+                  name: "ContextCompletionPrompt"
+                },
+                argument: {
+                  name: "value",
+                  value: "c"
+                },
+                context: {
+                  arguments: {
+                    locale: "en"
                   }
                 }
-              })
-              const result = yield* test.decodeResult(response).pipe(
-                Effect.flatMap((message) => decodeCompletion(message.result))
-              )
+              }
+            })
+            const result = yield* test.decodeResult(response).pipe(
+              Effect.flatMap((message) => decodeCompletion(message.result))
+            )
 
-              assert.deepStrictEqual(result.completion.values, ["context received"])
-            }))
-        }
+            assert.deepStrictEqual(result.completion.values, ["context received"])
+          }))
         // FIX: Effect currently returns a successful empty completion result for an unknown
         // prompt reference instead of rejecting the invalid reference.
         it.effect.skip("SHOULD reject an unknown prompt reference with Invalid Params", () =>
           Effect.gen(function*() {
-            const test = yield* McpConformanceTest
+            const test = yield* McpConformance
             const initialized = yield* test.initialize({ server: "features" })
             yield* test.notifyInitialized(initialized)
             const response = yield* test.send(initialized, {
@@ -126,7 +122,7 @@ export const suite = (protocol: McpProtocol.ProtocolAdapter, layer: TestLayer) =
         // argument name instead of rejecting the invalid input.
         it.effect.skip("MUST reject an unknown argument name", () =>
           Effect.gen(function*() {
-            const test = yield* McpConformanceTest
+            const test = yield* McpConformance
             const initialized = yield* test.initialize({ server: "features" })
             yield* test.notifyInitialized(initialized)
             const response = yield* test.send(initialized, {
