+++ packages/effect/test/unstable/ai/McpServer/McpServer.test.ts
@@ -20,7 +20,8 @@ import * as HttpClientRequest from "effect/unstable/http/HttpClientRequest"
 import * as HttpRouter from "effect/unstable/http/HttpRouter"
 import { RpcSerialization } from "effect/unstable/rpc"
 import * as RpcClient from "effect/unstable/rpc/RpcClient"
-import { makeHttpHarness, makeServerLayer } from "./McpServerTest.ts"
+import { makeHttpHarness } from "./TestUtils/McpHttpHarness.ts"
+import { makeServerLayer } from "./TestUtils/McpServerLayer.ts"
 
 const OptionalStringTool = Tool.make("OptionalStringTool", {
   parameters: Schema.Struct({ signature: Schema.optional(Schema.String) }),
