+++ packages/effect/test/unstable/ai/McpServer/TestUtils/McpHttpHarness.ts
@@ -1,45 +1,8 @@
 import * as Effect from "effect/Effect"
-import { constVoid } from "effect/Function"
 import * as Layer from "effect/Layer"
-import * as Logger from "effect/Logger"
-import * as References from "effect/References"
-import * as McpProtocol from "effect/unstable/ai/McpProtocol"
-import * as McpServer from "effect/unstable/ai/McpServer"
 import * as HttpRouter from "effect/unstable/http/HttpRouter"
 
 export const MCP_ENDPOINT = "http://localhost/mcp"
-
-const noopLogger = Logger.make(constVoid)
-
-export const makeServerLayer = (options: {
-  readonly name: string
-  readonly version?: string | undefined
-  readonly protocols?:
-    | readonly [
-      McpProtocol.ProtocolAdapter,
-      ...Array<McpProtocol.ProtocolAdapter>
-    ]
-    | undefined
-  readonly extensions?: Record<`${string}/${string}`, unknown> | undefined
-  readonly allowedOrigins?: ReadonlyArray<string> | undefined
-}) =>
-  McpServer.layerHttp({
-    name: options.name,
-    version: options.version ?? "1.0.0",
-    path: "/mcp",
-    protocols: options.protocols ?? [
-      McpProtocol.v2025_06_18,
-      McpProtocol.v2025_03_26,
-      McpProtocol.v2024_11_05
-    ],
-    extensions: options.extensions,
-    allowedOrigins: options.allowedOrigins
-  }).pipe(
-    Layer.provideMerge(Layer.succeed(
-      References.CurrentLoggers,
-      new Set([noopLogger])
-    ))
-  )
 
 export const makeHttpHarness = Effect.fnUntraced(function*<A, E>(
   serverLayer: Layer.Layer<A, E, HttpRouter.HttpRouter>,
